Add gnuplot to jepsen container
Also optimize re-running with the purge param.
Default false, which is remains running.

+++ vagrant_script/lein_test.sh
@@ -2,6 +2,7 @@
 # Pull images to new location, which is the shared docker volume /jepsen
 # Launch lein to test a given app ($1)
 # Protect from an incident running on hosts which aren't n1, n2, etc.
+# Stop/remove the main jepsen container, if $2 = purge
 hostname | grep -q "^n[0-9]\+"
 [ $? -eq 0 ] || exit 1
 [ "$1" ] || exit 1
@@ -18,24 +19,28 @@ then
   docker pull pandeiro/lein
 fi
 
+if [ "${2}" = "purge" ]; then
+  docker stop jepsen && docker rm -f -v jepsen
+fi
+
 # Run lein to make a custom galera dependency build
-docker stop jepsen && docker rm -f -v jepsen
+docker stop jepsen-build && docker rm -f -v jepsen-build
 echo "Make a custom galera jar build"
-docker run -it --rm \
+docker run -itd --rm \
   -v /jepsen/jepsen/galera:/app \
   --entrypoint /bin/bash \
-  --name jepsen -h jepsen \
+  --name jepsen-build -h jepsen \
   pandeiro/lein:latest -c "lein deps && lein compile && lein uberjar; sync"
 sync
 
 # FIXME(bogdando) remove those customs, when build is not required anymore
 # Run lein to make a custom jepsen build
-docker stop jepsen && docker rm -f -v jepsen
+docker stop jepsen-build && docker rm -f -v jepsen-build
 echo "Make a custom jepson jar build"
 docker run -it --rm \
   -v /jepsen/jepsen/jepsen:/app \
   --entrypoint /bin/bash \
-  --name jepsen -h jepsen \
+  --name jepsen-build -h jepsen \
   pandeiro/lein:latest -c "lein deps && lein compile && lein uberjar; sync"
 sync
 
@@ -52,6 +57,12 @@ docker run --stop-signal=SIGKILL -itd \
   --entrypoint /bin/bash \
   --name jepsen -h jepsen \
   pandeiro/lein:latest
+if [ "${2}" = "purge" ]; then
+  # install dependency
+  docker exec -it jepsen bash -c "apt-get update"
+  docker exec -it jepsen bash -c "apt-get -y install gnuplot-qt"
+fi
+# copy custom jar builds
 docker exec -it jepsen bash -c "mkdir -p resources/jepsen/galera/0.1.0-SNAPSHOT"
 docker exec -it jepsen bash -c "cp -f /custom2/jepsen.galera-0.1.0-SNAPSHOT*  resources/jepsen/galera/0.1.0-SNAPSHOT/"
 docker exec -it jepsen bash -c "mkdir -p resources/jepsen/jepsen/0.1.0-SNAPSHOT"
